Don't infinitely recurse when encountering a pointer.
Pointers are wrapped in a class object when reflected since they cannot
be boxed as an object directly.
Since this object has a pointer field that is boxed as a clone of the object,
it looks like an infinite linked list when reflected.
Unwrap the pointer and treat it like it is a primitive value.